[T4] Fix unit test failing on Windows
On Windows using a full path when creating an AssemblyName throws a
FileLoadException - 'The given assembly name or codebase was invalid.'
When resolving assembly references if the path is a full path then
the DummyTemplateGenerator now just returns the path. Fixing this then
caused the template code generation to fail with two compiler errors:

warning CS2023: Ignoring /noconfig option because it was specified in
a response file
error CS1703: An assembly with the same identity 'System,
Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089' has
already been imported. Try removing one of the duplicate references.

By removing the .dll part from the assembly reference the System.dll
is now correctly resolved so no duplicate System assembly is added.
This fixes the compiler errors.