MonoTargetRuntime: Use correct xbuild dirs for toolsVersion 12.0/14.0 or
.. fallback to `$prefix/lib/mono/4.5` if toolsVersion==4.0 .

This restores the older behavior from before the msbuild switch was
added.

Note: We only have a v4.0 builder for non-windows platforms, so even
though we may find `.../xbuild/12.0/xbuild.exe`, we will still fall back
to a 4.0 builder. To fix this, we need to add a 12.0 and 14.0 builder
which works against the xbuild assemblies.