[NuGet] Handle install error when NuGet package already installed
Fixed bug #42807 - Adding all the Google Play Services packages to an
Android app throws an error as some dependents packages already exist
https://bugzilla.xamarin.com/show_bug.cgi?id=42807

NuGet v3 has different behaviour to NuGet v2. If a NuGet package is
already installed then an attempt is made to install the same
package version again then NuGet v3 will throw a
PackageAlreadyInstalledException. NuGet v2 would log a message and
not throw an error.

The PackageAlreadyInstalledException is now caught and a message is
logged instead of having the exception thrown. This makes the
behaviour consistent with previous versions of the IDE.