[Ide] Add apply all check box to dialog when overwriting project files
When adding or moving multiple files into a project, and the files
already exist, the user will be shown a dialog asking if they want
to overwrite each file. The dialog has a Cancel, Skip and Overwrite
File button and an Apply to all check box when multiple files are
being added. If a single file is being added then the Skip and
Apply to all check box is not displayed.

Previously there was no way to cancel multiple files being added
or moved. The cancel button would only cancel the file currently
being added. Also there was no way apply the same action, skip or
overwrite, to all files without having to repeatedly click a button
on the dialog.