[Ide] Allow existing project files to be overwritten when adding files (#1535)
* [Ide] Allow existing project files to be overwritten when adding files

Fixed bug #20204 - Adding existing items to a folder does not offer
overwrite option
https://bugzilla.xamarin.com/show_bug.cgi?id=20204

Adding existing files to a folder in a project would not allow the
existing files to be overwritten. The existing project files would
need to be deleted first. Now when adding an existing file to a folder
in a project where they already exist the user is prompted to
overwrite the existing file.

* [Ide] Add apply all check box to dialog when overwriting project files

When adding or moving multiple files into a project, and the files
already exist, the user will be shown a dialog asking if they want
to overwrite each file. The dialog has a Cancel, Skip and Overwrite
File button and an Apply to all check box when multiple files are
being added. If a single file is being added then the Skip and
Apply to all check box is not displayed.

Previously there was no way to cancel multiple files being added
or moved. The cancel button would only cancel the file currently
being added. Also there was no way apply the same action, skip or
overwrite, to all files without having to repeatedly click a button
on the dialog.