[Packaging] Support opening and building .nuproj project files
Added a PackagingProject which derives from DotNetProject and an
associated project flavor so .nuproj files can be loaded and
compiled.

Added a new CompileTarget of Package since this is the OutputType
used by .nuproj project files. Without this new CompileTarget it
is not possible to open a .nuproj file without re-implementing the
OnReadProjectHeader logic. This change will need to be reviewed to
see if there is a better way of doing this.