[NUnit] Don't create an appdomain for the NUnit test runner itself.
In the case the nunit.framework assembly is not referenced with localcopy=true, this means that the nunit.framework assembly would be probed by the test runner inside the test directory, and it would not find nunit.framework.dll.

By not using an appdomain, we use the one which is discovered by SupportAssemblies via XS and then it can be run.

Bug 41921 - NUnit2 and NUnit3 testing Internal Error