[Packaging] Add basic project template wizard for .nuproj projects
Added a simple project template wizard that allows the package id,
version, authors and description to be specified. The wizard page
needs an image on the right hand side, there is nothing displayed
currently. There is no validation of the data entered apart from
checking it exists. Currently using a TextView for the description
to allow multi-lines to be specified which prevents the enter/return
key from moving to the next page of the dialog when the TextView
has focus.

Had to add the packaging project to the MonoDevelop.Ide project's
InternalsVisibleTo. Without this the build was failing to find
the GTK# designers BinContainer class even though it is in the
project's generated.cs file.