[ConnectedServices] Rework some api
Dependencies are (at present) one of two kinds; a nuget or some code that will
be added to the project. These need to be added to the project when the
service is added - they are, after all, _dependencies_.

The idea of a configuration section being added by the user is still useful
I think, but not so much for dependencies.