[NuGet] Allow a NuGet package and its dependencies to be removed
Added a new UninstallPackages method to the public API for the NuGet
addin that takes a project, a list of packages ids, and a flag that
can be used to optionally remove the dependencies of the NuGet
package has. With this flag set to true then uninstalling a NuGet
package such as jQuery.Validation will attempt to uninstall jQuery
too, since jQuery.Validation depends on it.