[Ide][GettingStarted] No Getting Started View restriction
Getting Started providers can decide how to show
the information to the user. A default implementation
to show a document tab can be used through
GettingStartedProvider base, which requires only
a custom widget.

This affects only the Getting Started Node, the provider
must still provide a single widget which could be
embedded in other places.