[Packaging] Output from referenced project now added to NuGet package
A normal .NET project can now be referenced by a packaging project
(.nuproj) and when the packaging project is built the assembly from
the referenced project will be included in the generated NuGet
package. Currently the packaging MSBuild targets are not global so
the referenced project is modified so it imports these packaging
targets. Removing the project reference will remove the imports from
the .NET project.