[Core] Fix C# formatting policy not saved for solution
Fixed bug #43246 - Custom code formatting (Indentation) settings are
not stored and reset after the solution restart.
https://bugzilla.xamarin.com/show_bug.cgi?id=43246

When a single C# formatting option was changed so it matched the
inherited policy then the solution was not saved since the policy
information in the solution was not being updated correctly.
When a setting  matches the inherited policy it is removed from the
properties that are to be saved with the solution file. However if
that property was originally stored in the solution file it would not
be removed when the solution was saved. A simple way to reproduce this
is to create a new solution, open the Solution Options dialog, select
Source Code - Code Formatting - C# source code, C# format, Edit,
then change Indent case sections in the Indentation category, close
and re-open the solution. If this repeated a few times then eventually
the option will not have the correct value. At this point changing
the option will not update the solution file at all.