[Core] Fix unknown custom solution section data being removed
The change to fix bug #43246 broke the KeepUnknownCustomData test
in the SolutionTests. So a different approach has been taken. Now
the formatting policy information is not removed when serializing
the data. Instead values which are the same as the inherited policy
have their DataNode's IsDefaultValue set to true. When saving the
solution section to a file these items with IsDefaultValue are
not saved.

Also added a test to ensure only those formatting properties which
are different to the inherited policy are saved in the solution file
if the solution does not have any policy information stored. The
DataItem's UpdateFromItem method was changed to ensure that child data
for a DataItem are not added if they are using the default values.
Without this change the formatting policy would just add all values
if the solution file did not have a formatting policy already saved.