[Packaging] Add MSBuild imports to .NET project on adding metadata
The common NuGet packaging MSBuild imports are not yet global so
in order to be able to create a NuGet package from a .NET project
these imports need to be added. If the NuGet package metadata in
project options has any data added then these imports will be added
to the project.