[ConnectedServices] Added some simple retry logic when applying code changes
This works around an issue where, after adding a reference to a project,
the TypeSystem updates the roslyn workspaces and the 'workspaceVersion' changes.
When this version changes, TryApplyChanges will return false and we miss
out on adding the code.

We check for this specific case and trigger a retry. Maximum attempts is 3.