[Packaging] Make target directory read only in Properties window
The target directory is now taken from the LogicalName metadata
which is stored on the ProjectFile's ResourceId property. When a
NuGetFile is displayed in the Properties window the Target property
shows the target directory for the file. This is determined from
either the link, the file path, or the ResourceId depending on what
is set. The Target property is now read only since the ResourceId
can be set instead and having both writable can cause one value to
overwrite the other. Also ResourceId includes the filename whilst
Target is the directory so they show different information.