[Stub] Implement support for hijacking language if none is set in XS.
This issue is new in Sierra, where the preferred user interface
language contains regional information (which it should not).

In El Capitan, having Language set to English and Region set to Romania
would set [NSLocale currentLocale] to en_RO, while [NSLocale
preferredLanguage[0]] to en.

In Sierra, both now happen to be en_RO, which causes issues if you have
multiple preferred languages and as a result, causes bug
https://bugzilla.xamarin.com/show_bug.cgi?id=43806