Use a higher limit for max threads
XS uses a lot of tasks/threads. To avoid deadlocking the IDE we should not
use such a low number. Mono's default is 400/400 at the time of writing. If
we actually want to limit it then lets go for a max of 125 threads. That gives
us plenty to work with, even if we do kick off 16 parallel nuget restores.

https://bugzilla.xamarin.com/show_bug.cgi?id=42342