[NuGet] Fix package incorrectly shown as not installed for app service
On the App Service page when the NuGet packages
Microsoft.Azure.Mobile.Client.SQLiteStore and
Microsoft.Azure.Mobile.Client were not installed then when the
SQLiteStore NuGet package was installed the
Microsoft.Azure.Mobile.Client NuGet package would still show as not
installed on the App Service page even though it was installed as a
dependency of the SQLiteStore NuGet package.

If the IPackageManagementProjectOperations GetInstalledPackages is
called after the PackageReferenceAdded event is fired the package
just installed was not available in the returned list of installed
packages. This is because NuGet has two events for packages. One event
is fired when a package is installed in the packages folder and
another is fired when a package is referenced by a project. The NuGet
addin was using the former which fires before the packages.config
file is updated so the package installed was not included in the list
of packages returned from GetInstalledPackages. To fix this the
PackageEvents PackageReferenceAdded event is used which is fired after
the packages.config file has been updated.