[NuGet] Optionally ignore missing package being uninstalled
When using the public API provided by the NuGet addin if a NuGet
package is missing when an attempt is made to uninstall it then an
error is no longer reported in the status bar. Instead a message will
be logged to the Package Console saying that the NuGet package does
not exist in the project.