[Packaging] Add platform implementation for Android
Right clicking a PCL project that has NuGet package metadata shows
an Add Platform Implementation menu. When this is selected a dialog
is displayed allowing platform implementations to be selected (iOS
and Android). Selecting Android and clicking the OK button will
add an Android project and NuGet packaging project to the solution
as well as moving the NuGet package metadata from the PCL project to
the NuGet packaging project.