[Packaging] Do not create platform projects outside solution directory
If the PCL project is in the same directory as the solution then the
platform implementation projects were being created outside the
solution's directory. Now the projects will be created inside the
solution in their own directories if the PCL project does not have
its own directory. If the PCL project does have its own directory
then the platform implementation projects are created inside the
parent directory of the PCL project.