[Packaging] Update NuGet package project template wizard
Fix empty author text box allowing the project to be created. The
author text box needs to be filled in.

Rename Authors label text to Author.

The description is now a single line text box instead of a multiline
text box. This makes it easier to create a new NuGet package since
pressing return when the description text box is focused will now
move to the next page in the New Project dialog. A multiline
description can be added in the project options dialog.