[Core] Fix project reference shown as incompatible when it is not
The DotNetProjectExtension has a OnGetCanReferenceProject method
which is used in the Edit References dialog to show compatible
projects but is not used in the solution pad in the References
folder. The ProjectReference has been changed to use the
DotNetProject's CanReferenceProject method which will call the
OnGetCanReference method which does the same check by default that
the ProjectReference was doing. This removes some code duplication
and also allows a DotNetProjectExtension to override the default
behaviour and specify which projects it supports. Previously even
though the DotNetProjectExtension indicated the project was
compatible the References folder would show the reference as being
incompatible.