[Editor] Ensure the layout cache sanity
Due to no guarantee in the order of which widgets are destroyed, we cannot
decide in all cases that the disposal of the layout cache happens at the
right time.

After the LayoutCache was disposed, more items were enqueued in the cache
as a result of disposing a TextViewMargin's layout cache.

https://github.com/mono/monodevelop/blob/c2fd3ed0be51b21a02cb86a8e1a68655df5f3910/main/src/core/Mono.Texteditor/Mono.TextEditor/Gui/TextViewMargin.cs#L1049-L1056