[Mac] Add algorithm to prefer last selected selector item
Given a small main window or long run
configuration/emulator/configuration names, we end up with a problem in
layouting the NSToolbar such information vital to the user is not
visible on screen.

Prefer the last selected item for drawing.

Bug 43675 - New "run configurations" menu means we can't see which
emulator is selected