[Ide] Rearchitect ProjectSearchCategory to optimize filtering
This change optimizes the searchbar behaviour to not re-query the
typesystem when the pattern is almost the same.

Also, use a dictionary instead of filtering every time we iterate, the
cache is not refreshed as often as a search is done.

This reduces CPU relative time: 83 calls with 20000ms compared to 60 calls
with 6000ms.

Relative time spent is 240ms vs 100ms per call