[Core] Add Memoization infrastructure.
Use at your own risk. Not threadsafe, leaky implementation, this is for performance testing.
It can be made threadsafe, but this depends on the implementation of the function being memoized, as one result being recalculated might not be worth it for thread safety.
This does not leak if the delegate can be finalized, so just don't hold onto the memoized function forever.