[Core] Allow default namespace to be defined in project template
The default namespace for a shared assets project or a .NET project
can now be defined in the project template via the options:

  <Options DefaultNamespace="CustomNamespace" />

This can be used to override the default namespace for a project
which is the project name. The DefaultNamespace parameter also
supports parameters (e.g. ${ProjectName}) which will be expanded.
This makes it simpler to set the same namespace for a group of
projects.