[MSBuild] Use a static env var dictionary for MSBuild evaluation.
This dictionary will be long lived, but it'll always contain the same values, as we evaluate inside the IDE env.
Properties are re-evaluated over and over between contexts, so we end up with a lot of Environment.GetEnvironmentVariable calls.
This reduces 175MB of allocations done by GetPropertyValue to 0.1MB allocated. The dictionary will be long lived, so it'll be reused in future evaluations.