[Packaging] Support generating reference assemblies.
Updated the MSBuild tasks so that if the project has reference
assembly generation configured then these assemblies are generated
and added to the NuGet package at compile time. The MSBuild tasks
are branched from an old version and stored in a different
repository. Support for generating reference assemblies will need
to be added to the final MSBuild tasks that will be used.