[Packaging] Use non-wildcard assembly version in projects
When creating a new multiplatform project or adding platform
implementation the assembly version in the AssemblyInfo.cs file is
set to be 1.0.0.0 instead of the default 1.0.* used by the file
template. The wildcard version can generate a different assembly
version for each project in the solution depending on how long each
project build takes. Using a fixed version prevents the build
warning if reference assemblies are being generated for a bait and
switch NuGet package.