[Editor] Only trigger reparsing of a doc if it has been shown.
This forces so that document reparsing is queued until the document is shown.
When opening a solution that had many documents open, it would trigger reparses for each document a few times, causing high CPU usage and data being calculated for probably no good reason.
Only trigger it for document switching and for documents that had their content requested. This way, we only do typesystem parsing on documents that are actually in use.