[Core] Improve MSBuildProjectService.EscapeString performance
This improves the allocations and CPU time of EscapeString.

On a benchmark of 1000000 runs on the string "%a%b%c%d%", we have the old
variant using 2473ms and the new variant 1381ms.

The allocations also improved, where the old variant would allocate 441MB
in total, and the new variant would allocate 188MB.