[Perf] Improve enumerators in internal custom collections
This makes it so there is no Enumerator created, as the compiler is smart
enough to detect that if there is a strongly typed GetEnumerator() method
with a return value that structurally matches IEnumerator, it'll prefer
that.
In this case, we're talking about structs. This will save us IEnumerator
allocations all over the place.