[NuGet] Fix types unavailable when using a project.json file
When a NuGet package was added to a project that uses a project.json
the types from the assembly provided by the NuGet package were not
available until the project was reloaded or the references were
modified. So code completion would not work for the NuGet package
just added.

The problem was that the type system needs to be refreshed when the
project.json file is updated so the new assembly references are
used since they are not added directly to the project file. This is
now done by calling NotifyModified on the project indicating that
the References have changed. The type system then reloads the
project.