[Ide] Optimize MonoDevelopTextLoader file querying.
This improves speed by not doing IO check for existence, but relying on the exception being thrown.
Also, log an error on the file that's unable to be loaded and the exception on the filestream. This way we can probably look in case there's a bug.
The old code also had a race where the file could be deleted between the exists check and the filestream opening.