[NuGet] Prevent second restore being triggered by a package install
When a NuGet package is installed, removed or updated the
project.lock.json file is automatically restored during by NuGet.
After the NuGet addin has completed the package operation it
uses the FileService to notify that the project.json file has changed.
This was then triggering another check in the NuGet addin's startup
handler to see if a package restore should be run when there was no
need to do so.

Now a check is made to see if the NuGet addin is currently running
a package operation and if so the package restore check is not done.
This change also fixes the project.json file not being refreshed in
the text editor.

Directly editing the project.json file in the text editor will still
trigger the package restore if the lock file has changed.