[Ide] Fix searchbar crashing on being typed in
There was no information in the logs about this, so all the lead I had was this stacktrace:
WARNING [2016-10-10 13:58:25Z]: Highlighting Theme Dark not found, using default Dark instead
WARNING [2016-10-10 13:58:25Z]: Color style Dark not found, switching to default.
ERROR [2016-10-10 13:58:25Z]: Color style Dark not found.
FATAL ERROR [2016-10-10 13:58:25Z]: Xamarin Studio failed to start. Some of the assemblies required to run Xamarin Studio (for example gtk-sharp)may not be properly installed in the GAC.
System.NullReferenceException: Object reference not set to an instance of an object
  at (wrapper managed-to-native) Gtk.Application:gtk_main ()
  at Gtk.Application.Run () [0x00001] in /private/tmp/source-mono-4.8.0/bockbuild-mono-4.8.0-branch/profiles/mono-mac-xamarin/build-root/gtk-sharp-None/gtk/Application.cs:135
  at MonoDevelop.Ide.IdeApp.Run () [0x00001] in /Users/therzok/Work/md/monodevelop/main/src/core/MonoDevelop.Ide/MonoDevelop.Ide/Ide.cs:386
  at MonoDevelop.Ide.IdeStartup.Run (MonoDevelop.Ide.MonoDevelopOptions options) [0x0079c] in /Users/therzok/Work/md/monodevelop/main/src/core/MonoDevelop.Ide/MonoDevelop.Ide/IdeStartup.cs:283
  at MonoDevelop.Ide.IdeStartup.Main (System.String[] args, MonoDevelop.Ide.Extensions.IdeCustomizer customizer) [0x000ae] in /Users/therzok/Work/md/monodevelop/main/src/core/MonoDevelop.Ide/MonoDevelop.Ide/IdeStartup.cs:594
After bisecting, I pinned it down to the bit of code in this commit which was removed in b9a8cac6a94194d2698767faac81d3d5874271fe