[Perf] Optimize ToArray where applicable.
ToArray is slightly less efficient than ToList, as it does one extra allocation to reserve _exact_ size for the collection.
Switch to ToList on stuff which came up on the profiler and is as simple as re-iterating the same collection after toarray/tolist.
The change in OnGetRunConfigurations is a bit more complex, as we used to create two linqs for something that could've already been evaluated. The code is now more efficient in all regards.