[Packaging] Allow IncludeInPackage to be set for files in properties
The IncludeInPackage metadata is added to project file items to
have them included in the NuGet package. The IncludeInPackage
property can now be set in the Properties window.

<None Include="readme.txt">
  <IncludeInPackage>true</IncludeInPackage>
</None>

Currently when setting the value back to false, which is the default,
the value is not removed from the project file. There seems to be a
problem in the project model where the value is removed in memory
but not removed when saving the project to a file.