[Packaging] Fix Create NuGet Package not creating a .nupkg
Setting the PackOnBuild to true as a global MSBuild property does not
work with MSBuild but does with XBuild. The property does not seem to
have any affect on properties in the project or imported. As a
workaround the Pack target is being called directly. Also still using
the IdeApp.ProjectOperations.Build method so build errors are reported
in the Errors window. Running the Pack target directly on the project
without using ProjectOperations.Build results in no information being
reported in the IDE.