[NuGet] Provide a way to install NuGet packages in unit tests
Project templates that have NuGet packages can now be installed
in unit tests via the NuGetPackageInstaller class which is available
in the NuGet addin tests project. This class has an InstallPackages
method which takes a solution and the list of NuGet packages
returned from the project template after the main project is created.