[NuGet] Support package restore for .NET Core projects
Currently the NuGet addin is running dotnet restore3 to restore
NuGet packages. At some point this should change and instead NuGet's
DependencyGraphRestoreUtility class should be used. This requires
support in MSBuild to provide the dependency graph information.

If automatic package restore is enabled in preferences then a restore
will always happen for .NET Core projects. This will need to be
looked at in the future since a restore should only happen if NuGet
packages are missing or if generated .targets files are missing.