[NuGet] Fix status bar warning when installing same NuGet package
When installing the same version of a NuGet package that already
exists in a .NET Core project a warning was reported in the status
bar. This is inconsistent with the behaviour for other projects where
they show a success message in the status bar and report that the
NuGet package already exists in the Package Console window.