[Packaging] Fix NuGet projects not building referenced projects
When using the Create NuGet Package command on a NuGet packaging
project (.nuproj) the referenced projects would not be built. One
way to reproduce this is to create a new multiplatform project for
iOS and Android and then use the Create NuGet Package command on the
NuGet packaging project. The build will fail since the iOS and Android
projects are not compiled. This is because the Pack MSBuild target
is used which will only build the project it is being run on. Now
for NuGet packaging projects when the Create NuGet Package command
is selected a normal build is run for the project which ensures
referenced projects are built.