[NuGet] Fix types unavailable from NuGet package in .NET Core project
After adding a NuGet package to a .NET Core project the type
information would not be available in the text editor for code
completion. This was because the reference information was not
being refreshed after the NuGet package is installed. Now when
a NuGet package restore occurs for the .NET Core project then the
type system is notified it needs to reload the project. This is
done by calling the project's NotifyModified method passing in the
References string.