[DotNetCore] Fix being unable to run app after restore
After restoring the NuGet packages for a .NET Core project extra
MSBuild targets files are generated in the obj directory. If the
project is then re-opened the output path used to run the application
was incorrect. This was because when the extra MSBuild targets files
are available on project load the configuration's OutputDirectory
is correctly set and the target framework was being combined with it
producing an incorrect output directory. Now if the configuration's
OutputDirectory is an absolute path it is used otherwise the
original logic is used which works when a project is first opened
and the MSBuild targets files are not available.