[DotNetCore] Add status bar information when creating new project
Creating the project from a template happens after the New Project
dialog has closed and the solution has opened. Since it takes a
bit of time to do this it looks like nothing is happening. So now
a status bar message is shown indicating that the project is being
created.