[Packaging] Copy NuGet.Build.Packaging package to cache if missing
To make sure a NuGet packaging project can be built the
NuGet.Build.Packaging package needs to be installed into the global
machine cache. The first template the NuGet package project
template is installed a check is made to see if the package is
available in the global cache and if not it is installed there.