[Ide] Fix preselection of grouped templates in new project dialog
SelectTemplate(templateId) does not take into account
that a template can be part of a group, where the parent
category references only the first template.

This patch adds an additional check for template group
members. If a group contains the requested template,
it selects the group in the dialog and the target language
accordingly.

(fixes bug #46332)