[NuGet] Allow project templates to specify package directory
Each package reference in a project template can now specify a
directory relative to the current directory where a NuGet package
will be available. Previously an addin could only define the directory
through an extension which meant that it would be used for any
project template that had package references. This was fine for
existing projects that used packages.config or project.json files
but for newer project types that implement the INuGetAwareProject
interface this causes a problem since NuGet will arbitrarily pick
the first package source even though that source may not have the
NuGet package. NuGet will then fail when it tries to execute the
project actions since the package is unavailable from the source
associated with the action. Having an explicit directory for each
package reference in a project template avoids this since only one
package source is used in this case and the correct one is
associated with the package action returned from the
NuGetPackageManager's PreviewInstallPackageAsync method.