[Packaging] Fix error installing NuGet package into nuproj project
The project templates now explicitly define the directory where the
NuGet.Build.Packaging NuGet package will exist. This prevents the
install failing in NuGet packages since NuGet picks the first
package source from the primary sources, which in this case would
be all the local directories having NuGet packages defined by the
various addins, which may not have the NuGet package. NuGet would
then fail when it attempted to execute the package action since the
package source would not have the NuGet package. Using an explicit
package directory in the project templates for each package reference
avoids this problem since only that directory will be used as a
package source.