[Core] Optimize the generation of Execution parameters.
This commit introduces the caching of GetCustomAttribute calls on the MonoExecutionParameters type.
This is actually an invariant across all the calls.
The actual slowness appeared on the UI thread update timer, as the option generation is called for the Run button update.
This removes 87ms on 53 calls from the command update timer.