[MSBuild] Optimize Transformed includes evaluation.
This PR introduces a change in how transformed includes are handled.

In the previous case, we would try to evaluate the metadata properties on
a transformed include, even in cases where we would have properties which
could not be evaluated.

<ResxCode Include="@(_LinkedResx -> WithMetadataValue('Generator', 'PublicResXFileCodeGenerator'))">
	<Public>True</Public>
	<CanonicalRelativeDir>$([System.IO.Path]::GetDirectoryName('%(Link)'))</CanonicalRelativeDir>
</ResxCode>

This would cause the MSBuild parser to try and evaluate the %(Link)
property on an item which does not have any evaluated metadata (the
transforms aren't parsed until later in the code which does a check for
whether a MSBuild item is a transform).

This optimizes evaluation so it skips metadata evaluation on transform
items.

Fixes log spew in md-addins

Bug 43509 - Failure to evaluate msbuild item properties