Fix dynamic type keyword syntax highlighting
The dynamic keyword is not highlighted properly. In the UI, you can see
that it is actually highlighted as a keyword, but is quickly (< 1 second)
reverted to the default text color.

Removing "dynamic" from the list fixes the problem. There is a comment
that suggests that "var" and "dynamic" should be skipped here. "var" is
already omitted from the list.