[NuGet] Support opening Add Packages dialog from Dependencies
The Add Packages dialog can now be opened by right clicking the
Dependencies folder or the NuGet folder. It can also be opened by
double clicking the NuGet folder.

Also added the Restore menu for restoring packages which is available
by right clicking the Dependencies folder or the NuGet folder.