[DotNetCore] Fix text editor warnings shown for new project
After creating a new .NET Core project and restoring the NuGet
packages the project file is reloaded to ensure the correct
configuration information is available to remove the warnings in the
text editor. This also fixes the wrong target framework for the
project which can be seen when trying to reference a .NET Core
project in the Edit References dialog.