[Packaging] Fix error creating NuGet packaging project on Windows
Fixed bug #47169 - Unable to create NuGet package on Windows
https://bugzilla.xamarin.com/show_bug.cgi?id=47169

Creating new NuGet packaging project after just opening the IDE would
report an error saving the solution due to the packaging addin not
being able to find the NuGet.Packaging.Core.Types. This problem also
prevented NuGet package metadata from being saved into the NuGet
packaging project (.nuproj). Without the metadata a NuGet package
(.nupkg) would not be created when building the project.

The NuGet.Packaging.Core.Types.dll is included with the NuGet addin
and is now imported by that addin so it is available for the packaging
addin to use.