[Packaging] Add the build NuGet package when project is referenced
Fixed bug #47372 - NuGet Package template looks confusing
https://bugzilla.xamarin.com/show_bug.cgi?id=47372

The original problem was that a NuGet packaging project and a class
library project was created. The library project had a NuGet package
dependency. The library project was then referenced by the NuGet
packaging project and on building the NuGet package it contained
the library assembly and the NuGet package dependency assemblies
in its lib directory. It did not have the NuGet package as a
dependency in the .nuspec file. This is because the library project
needs the NuGet.Build.Packaging NuGet package installed for the
correct information to be added to the generated NuGet package. This
is confusing for a first time user.

Now when a reference to another project is added to a NuGet packaging
project then the NuGet.Build.Packaging NuGet package is installed
into that project if it is missing. This puts the referenced project
in the preferred state so that any NuGet package dependencies it has
are correctly added to the NuGet package generated by the NuGet
packaging project and only the library assembly is added to the lib
directory. Without the NuGet.Build.Packaging NuGet package
installed only primary outputs from the project can be collected
and package dependencies are ignored.