[NuGet] Fix package sources not shown when NuGet.Config read fails
Fixed bug #47419 - Exception reading NuGet configuration causes
sources panel to be blank
https://bugzilla.xamarin.com/show_bug.cgi?id=47419

If there is an error decrypting the passwords saved in the
global NuGet.Config file then the package sources in preferences
was empty. The exception thrown was not handled causing the UI
widget to not be created.

Now when there is an exception trying to read the NuGet.Config file
the package sources panel will show an error message and have a
link that will open the NuGet.Config file into the text editor.