[NuGet] Reload settings if NuGet.Config file is changed
Monitor changes made to any NuGet.Config file and if it has changed
then the NuGet settings for any open solution is reloaded. This
allows a NuGet.Config file to be edited in the text editor and
for changes to the package sources to be available the next time
the Add Packages dialog is opened. If the NuGet.Config file cannot
be read then an error dialog will be displayed in the same way
that an error dialog is shown on opening a solution if the
NuGet.Config file cannot be read.