[NuGet] Do not clear Package Console on creating project templates
When NuGet packages are installed when a new project is created the
Package Console is now not cleared. When a new solution is being
created the Package Console will be cleared since the log view
monitor automatically clears itself if a solution is closed or
opened. So only when a new project is added to an existing solution
will the Package Console is not cleared. This is to allow .NET Core
projects to run a restore during project creation as well as have
another project install NuGet packages and ensure that the NuGet
output from both is not lost in the Package Console window.